Separate run workflow for scse processing (#605)
* Refactor run_workflow out of cli

* Sort imports and formatting

* Fix type hint

* Move run_workflow to core.py

* Change input to run_workflow to accept dataframe
